Add files via upload
The Python programming language is a great tool to use when working with numbers and evaluating mathematical expressions. This quality can be utilized to make useful programs.
In this assignment,you will be learning how to make a calculator program in Python. This calculator will be able to perform various mathematical/arithmetic operations, but the final step of this guide serves as a starting point for how you might improve the code to create a more robust calculator.
The goal is to create a simple calculator which can perform basic arithmetic operations like addition, subtraction, multiplication or division depending upon the user input.